护石

+++ src/main/kotlin/com/linweiyuan/mhp/entity/User.kt
@@ -9,7 +9,7 @@ data class User(
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
         @JsonIgnore
-        val id: Int,
+        val id: Byte,
         val username: String,
         val password: String,
         @JsonIgnore
